template-service-broker: move pkg/template/servicebroker/apis/config to staging
