+++ staging/src/github.com/openshift/template-service-broker/apis/config/register.go
