+++ staging/src/github.com/openshift/template-service-broker/apis/config/v1/register.go
